cleaned up legacy files
