+++ src/tools/movieSearch.ts
@@ -1,4 +1,3 @@
-import type { ToolFn } from "../../types";
 import { z } from "zod";
 import { openai } from "../ai";
 import { queryMovies } from "../rag/query";
@@ -6,116 +5,6 @@ import type { MovieMetadata } from "../rag/types";
 import { styleWithPersona } from "../lib/persona-styler";
 import { PersonalityKey } from "../constants/personalities";
 
-export const movieSearchToolDefinition = {
-    name: "movie_search",
-    parameters: z.object({
-        query: z.string().describe(
-            "Search query for movies. Use specific terms like 'alien invasion', 'space exploration', 'action thriller', 'romantic comedy', 'highly rated sci-fi', 'movies from 1990s', etc. Be descriptive and specific."
-        ),
-        genre: z.string().nullable().describe('Filter movies by genre (e.g., "Action", "Comedy", "Sci-Fi"). Only use when user specifically mentions a genre by name.'),
-        director: z.string().nullable().describe('Filter movies by director name (e.g., "Christopher Nolan", "Steven Spielberg"). Only use when user specifically mentions a director by name.'),
-        year: z.number().nullable().describe('Filter movies by year (e.g., 1995, 2000). Only use when user specifically mentions a year or decade (e.g., "90s" = 1990s, "2000s" = 2000s).'),
-        limit: z.number().nullable().describe('Number of movies to return. Use 1 for single recommendations (like "the best", "ultimate", "top pick"). Use null or higher numbers for multiple recommendations.'),
-    }),
-    description: `ALWAYS use this tool when users ask about movies, want movie recommendations, suggestions, or ask to find specific movies. This tool searches a database of movies with metadata including title, year, genre, director, actors, rating, votes, revenue, metascore, and descriptions. Use for ANY movie-related query. 
-
-FILTERING GUIDELINES:
-- Use director filter when user mentions a specific director by name
-- Use genre filter when user mentions a specific genre by name  
-- Use year filter when user mentions a specific year (e.g., 1995) or decade (e.g., "90s" = 1990s, "2000s" = 2000s)
-- Use limit=1 when user asks for a single recommendation ("the best", "ultimate", "top pick", "recommend me one movie")
-- Use limit=null or higher when user asks for multiple recommendations ("recommendations", "suggestions", "list of movies")
-- The vector search is very effective at finding relevant movies based on descriptions and themes`,
-};
-
-type Args = z.infer<typeof movieSearchToolDefinition.parameters>;
-
-export const movieSearch: ToolFn<Args, string> = async ({ toolArgs, userMessage, personality }: { toolArgs: Args; userMessage: string; personality?: PersonalityKey; }) => {
-
-    const { query, genre, director, year, limit } = toolArgs;
-
-    const filters = {
-        ...(genre && genre !== null && { genre }),
-        ...(director && director !== null && { director }),
-        ...(year && year !== null && { year }),
-    };
-
-    // Determine how many results to return
-    const actualLimit = limit ?? 5; // Default to 5 if not specified
-    const isSingleRecommendation = actualLimit === 1;
-
-    console.log('filters', filters);
-    console.log('limit', actualLimit, 'isSingle', isSingleRecommendation);
-    console.log('toolArgs.limit:', limit);
-    console.log('personality (movieSearch):', personality);
-
-    let results;
-
-    try {
-        results = await queryMovies({
-            query: toolArgs.query,
-            filters,
-            topK: Math.max(actualLimit, 3)
-        });
-    } catch (e) {
-        console.error('Movie search error:', e);
-        return JSON.stringify({
-            type: 'movie_recommendations',
-            metadata: {
-                title: 'Movie Search Error',
-                description: 'Unable to search movies at this time'
-            },
-            data: {
-                recommendations: []
-            },
-            contextualMessage: 'Sorry, I encountered an error while searching for movies. Please try again later.'
-        });
-    }
-
-    const formattedResults = results.map((result: any) => {
-        const metadata = result.metadata as MovieMetadata;
-        return { ...metadata };
-    });
-
-    // Handle empty results
-    if (formattedResults.length === 0) {
-        return JSON.stringify({
-            type: 'movie_recommendations',
-            metadata: {
-                title: 'No Movies Found',
-                description: 'No movies match your search criteria'
-            },
-            data: {
-                recommendations: [],
-                query: query,
-                genre: genre,
-                year: year,
-                director: director
-            },
-            contextualMessage: `Sorry, I couldn't find any movies matching your search for "${query}". Note: My movie database contains films from 2006 onwards. If you're looking for older movies, try searching for recent films with similar themes or genres.`
-        });
-    }
-
-    // Return structured format for movie recommendations
-    const structuredResponse = {
-        type: 'movie_recommendations',
-        data: {
-            recommendations: formattedResults.slice(0, actualLimit), // Use the specified limit
-            query: query,
-            genre: genre
-        },
-        metadata: {
-            title: isSingleRecommendation ? 'Movie Recommendation' : 'Movie Recommendations',
-            description: isSingleRecommendation
-                ? `Found the perfect movie for "${query}"`
-                : `Found ${formattedResults.length} movie recommendations for "${query}"`
-        },
-        contextualMessage: await generateContextualMessage(formattedResults, query, genre, director, isSingleRecommendation, personality),
-        aiChosenMovie: isSingleRecommendation ? formattedResults[0] : await getAiChosenMovie(formattedResults, query, genre, director)
-    };
-
-    return JSON.stringify(structuredResponse, null, 2);
-};
 
 
 async function getAiChosenMovie(movies: any[], query: string, genre?: string | null, director?: string | null): Promise<any> {
